# apps/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 简浩影视 Admin API
  version: 1.0.0
  description: |
    后台管理系统 API，涵盖认证、用户、角色、项目、拍摄会话、预约、摄影师、服务、作品集、内容管理、审计日志与仪表盘统计，并提供对外官网数据接口。
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /users:
    get:
      summary: 用户列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: 用户列表
    post:
      summary: 创建用户
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName]
              properties:
                email: { type: string }
                password: { type: string }
                displayName: { type: string }
                roles:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: 已创建
...
  /roles:
    get:
      summary: 角色与权限列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 角色列表
    post:
      summary: 创建角色
      security:
        - bearerAuth: []
      responses:
        '201':
          description: 已创建
...
  /projects:
    get:
      summary: 项目列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 项目数据
    post:
      summary: 创建项目
      security:
        - bearerAuth: []
      responses:
        '201':
          description: 项目已创建
...
  /sessions:
    get:
      summary: 拍摄会话列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 会话数据
    post:
      summary: 创建拍摄会话
      security:
        - bearerAuth: []
      responses:
        '201':
          description: 会话已创建
...
  /audit-logs:
    get:
      summary: 审计日志
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 审计数据
...
  /bookings:
    get:
      summary: 预约列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: status
          schema:
            type: string
            description: 预约状态枚举值
        - in: query
          name: search
          schema:
            type: string
            description: 按客户姓名/邮箱模糊搜索
        - in: query
          name: serviceId
          schema:
            type: string
        - in: query
          name: photographerId
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 预约分页数据
    post:
      summary: 创建预约
      description: 需要 `bookings:manage` 权限。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientName, clientEmail]
              properties:
                clientName:
                  type: string
                clientEmail:
                  type: string
                  format: email
                clientPhone:
                  type: string
                eventDate:
                  type: string
                  format: date-time
                venue:
                  type: string
                status:
                  type: string
                note:
                  type: string
                serviceId:
                  type: string
                photographerId:
                  type: string
                accessCode:
                  type: string
                  description: 自定义 6 位访问口令，可选
      responses:
        '201':
          description: 预约已创建
...
  /photographers:
    get:
      summary: 摄影师列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: boolean
      responses:
        '200':
          description: 摄影师分页数据
    post:
      summary: 创建摄影师
      description: 需要 `photographers:manage` 权限。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, bio, specialties, experienceYears]
              properties:
                name:
                  type: string
                slug:
                  type: string
                avatarUrl:
                  type: string
                  format: uri
                bio:
                  type: string
                tagline:
                  type: string
                specialties:
                  type: array
                  items:
                    type: string
                experienceYears:
                  type: integer
                accolades:
                  type: array
                  items:
                    type: string
                socialLinks:
                  type: object
                isFeatured:
                  type: boolean
                isActive:
                  type: boolean
      responses:
        '201':
          description: 摄影师已创建
...
  /services:
    get:
      summary: 服务套餐列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: 服务分页数据
    post:
      summary: 新增服务套餐
      description: 需要 `services:manage` 权限。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, summary]
              properties:
                name:
                  type: string
                slug:
                  type: string
                summary:
                  type: string
                description:
                  type: string
                priceFrom:
                  type: number
                duration:
                  type: integer
                deliverables:
                  type: array
                  items:
                    type: string
                isPopular:
                  type: boolean
      responses:
        '201':
          description: 服务已创建
...
  /portfolio:
    get:
      summary: 作品集列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: 作品集分页数据
    post:
      summary: 新增作品
      description: 需要 `portfolio:manage` 权限。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, coverUrl]
              properties:
                title:
                  type: string
                coverUrl:
                  type: string
                  format: uri
                galleryUrls:
                  type: array
                  items:
                    type: string
                    format: uri
                category:
                  type: string
                description:
                  type: string
                featured:
                  type: boolean
      responses:
        '201':
          description: 作品已创建
...
  /public/bookings:
    post:
      summary: 官网预约提交
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientName, clientEmail]
              properties:
                clientName:
                  type: string
                clientEmail:
                  type: string
                  format: email
                clientPhone:
                  type: string
                eventDate:
                  type: string
                  format: date-time
                venue:
                  type: string
                serviceId:
                  type: string
                photographerId:
                  type: string
                note:
                  type: string
      responses:
        '201':
          description: 预约提交成功
...